requirements/requirements.txt
-r boilerplate.requirements.txt

# Add project-specific requirements here.
pillow==12.0.0
django-resized==1.0.3
